Fix and change
- Gave the hand type in the desc a little highlight
- Fixed it so that the message actually work (as intended?):

It shows correct level color for the hand type pops up. Before this fix the color of last hand type's level was shown, but the name was THIS hand.

Also I did a change, for example your last hand was High Card and this hand is Pair, the message shows `High Card !` then `Pair...`, meaning that the base mult and chips of your last hand got applied! and your current played hand is pending...